Update Problem_set_2.ipynb
We did all the calculations in a Mac M1 Pro using Python and VSCode. However, when we wanted to implement it in Google Collab, the second model gave a lot of problems. For now, we are updating the results. However, we want to finish the results in GC for you to see it.